separated high fidelity & original cad coloring as two containers
